+++ BloodFlow/BloodFlow.DataLayer/Entities/DonorOrder.cs
@@ -13,15 +13,15 @@ namespace BloodFlow.DataLayer.Entities
         // Overhiding the name of the variable from base entity and
         // set not mapped to not create this column in the database
         [NotMapped]
-        public new long Id { get; set; }
+        public new int Id { get; set; }
 
         [Column("donor_id")]
         [ForeignKey(nameof(Donor))]
-        public long DonorId { get; set; }
+        public int DonorId { get; set; }
 
         [Column("order_id")]
         [ForeignKey(nameof(Order))]
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
 
         public Donor Donor { get; set; } = null!;
 
@@ -29,9 +29,9 @@ namespace BloodFlow.DataLayer.Entities
 
         public DonorOrder() : base(0) { }
 
-        public DonorOrder(long id,
-            long donorId,
-            long orderId) : base(id)
+        public DonorOrder(int id,
+            int donorId,
+            int orderId) : base(id)
         {
             DonorId = donorId;
             OrderId = orderId;
