+++ BloodFlow/BloodFlow.DataLayer/Entities/PersonContact.cs
@@ -13,15 +13,15 @@ namespace BloodFlow.DataLayer.Entities
         // Overhiding the name of the variable from base entity and
         // set not mapped to not create this column in the database
         [NotMapped]
-        public new long Id { get; set; }
+        public new int Id { get; set; }
 
         [Column("peson_id")]
         [ForeignKey(nameof(Person))]
-        public long PersonId { get; set; }
+        public int PersonId { get; set; }
 
         [Column("contact_type_id")]
         [ForeignKey(nameof(ContactType))]
-        public long ContactTypeId { get; set; }
+        public int ContactTypeId { get; set; }
 
         [Column("contact_value")]
         public string ContactValue { get; set; } = null!;
@@ -32,9 +32,9 @@ namespace BloodFlow.DataLayer.Entities
 
         public PersonContact() : base(0) { }
 
-        public PersonContact(long id,
-            long personId,
-            long contactTypeId,
+        public PersonContact(int id,
+            int personId,
+            int contactTypeId,
             string contactValue) : base(id)
         {
             PersonId = personId;
